+++ app/internal/store/model.go
@@ -9,15 +9,15 @@ import (
 type Store struct {
 	bun.BaseModel `bun:"table:store,alias:s" swaggerignore:"true"`
 
-	StoreID   int64            `bun:"store_id,pk,autoincrement" json:"store_id"`
+	StoreID   int              `bun:"store_id,pk,autoincrement" json:"store_id"`
 	StoreName string           `bun:"store_name,notnull"        json:"store_name"`
-	AccountID int64            `bun:"account_id"                json:"account_id"`
+	AccountID int              `bun:"account_id"                json:"account_id"`
 	Account   *account.Account `bun:"rel:belongs-to,join:account_id=account_id" json:"account,omitempty"`
 }
 
 type CreateStoreInput struct {
 	StoreName string `json:"store_name" binding:"required"`
-	AccountID int64  `json:"account_id" binding:"required"`
+	AccountID int    `json:"account_id" binding:"required"`
 }
 
 type UpdateStoreInput struct {
